Make ROM reload robust enough
Big thing here is to check uart rx before switching to start bit state when we see the high-low transition and wait half a clock period. This ensures that spurious rx lows are ignored (including the reset weirdness I saw earlier), and makes the thing a bit more stable it looks like. Still fails sometimes, but a re-upload seems to fix it.

Still some more work on the loader; I'd love for it to watch the filesystem for rom rebuilds and make sure it prints any uart bytes sent from the xenowing, but for now this is good enough to merge I think. Also need to catch up on some of the simulator things.